fix(test_longevity): Rewrite the test to not use pytester
* Change config to test-cases/features/elasticity/longevity-elasticity-many-small-tables.yaml
Fixes #11625
